+ Map.js разбил на функции, инициализацию карты вынес в отдельный модуль
